+++ app/src/main/java/com/example/eventmanageri/Activities/EventModActivity.java
@@ -1,16 +1,19 @@
-package com.example.eventmanageri;
+package com.example.eventmanageri.Activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.eventmanageri.Database;
+import com.example.eventmanageri.Models.Event;
+import com.example.eventmanageri.R;
 
 import java.util.List;
