+++ app/src/main/java/com/example/eventmanageri/Adapters/RecyclerView_Config.java
@@ -1,7 +1,6 @@
-package com.example.eventmanageri;
+package com.example.eventmanageri.Adapters;
 
 import android.content.Context;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,6 +18,9 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.eventmanageri.Activities.EventViewActivity;
+import com.example.eventmanageri.Models.Event;
+import com.example.eventmanageri.R;
 import com.google.firebase.auth.FirebaseAuth;
