+++ app/src/main/java/com/example/eventmanageri/Models/Event.java
@@ -1,4 +1,4 @@
-package com.example.eventmanageri;
+package com.example.eventmanageri.Models;
 
 // Define items
