+++ app/src/main/java/com/example/eventmanageri/Models/User.java
@@ -1,4 +1,4 @@
-package com.example.eventmanageri;
+package com.example.eventmanageri.Models;
 
 public class User {
     private String key;
