+++ app/src/main/res/layout/activity_event_mod.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventModActivity">
+    tools:context=".Activities.EventModActivity">
 
     <Spinner
         android:id="@+id/share_sp"
