+++ app/src/main/res/layout/activity_new_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewEventActivity">
+    tools:context=".Activities.NewEventActivity">
 
     <Spinner
         android:id="@+id/share_sp"
