+++ app/src/main/res/layout/activity_test.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TestActivity">
+    tools:context=".Activities.TestActivity">
 
     <Button
         android:id="@+id/sv"
